+++ include/eolspacing.h
@@ -8,26 +8,26 @@ namespace phydb {
 class EolSpacing {
   private:
     //metal layer
-    float spacing_;
-    float eol_width_;
-    float eol_within_;
-    float par_edge_;
-    float par_within_;
+    double spacing_;
+    double eol_width_;
+    double eol_within_;
+    double par_edge_;
+    double par_within_;
 
   public:
     EolSpacing() : spacing_(0), eol_width_(0), eol_within_(0),
                    par_edge_(0), par_within_(0) {}
 
     //constructor for metal layer
-    EolSpacing(float spacing, float eolWidth, float eolWithin,
-               float parEdge, float parWithin) : 
+    EolSpacing(double spacing, double eolWidth, double eolWithin,
+               double parEdge, double parWithin) :
         spacing_(spacing), 
         eol_width_(eolWidth),
         eol_within_(eolWithin), 
         par_edge_(parEdge), 
         par_within_(parWithin) {}
     
-    float GetSpacing() {return spacing_;}
+    double GetSpacing() {return spacing_;}
 
     void Reset();
 };
