+++ include/site.h
@@ -9,23 +9,23 @@ class Site {
   private:
     string name_;
     string class_name_;
-    float width_;
-    float height_;
+    double width_;
+    double height_;
 
   public:
     Site() : name_(""), class_name_(""), width_(0), height_(0) {}
-    Site(string name, string className, float width, float height) :
+    Site(string name, string className, double width, double height) :
         name_(name), class_name_(className), width_(width), height_(height) {}
 
     void SetName(string);
     void SetClassName(string);
-    void SetWidth(float);
-    void SetHeight(float);
+    void SetWidth(double);
+    void SetHeight(double);
 
     string GetName() const;
     string GetClassName() const;
-    float GetWidth() const;
-    float GetHeight() const;
+    double GetWidth() const;
+    double GetHeight() const;
 
 };
